requirements.txt
yt-dlp==2025.11.12
requests>=2.31.0
pyyaml>=6.0